+++ lib/src/enums/ad_network.dart
@@ -1,12 +1,6 @@
 import 'package:flutter/foundation.dart';
 
-enum AdNetwork {
-  any,
-  admob,
-  facebook,
-  appLovin,
-  unity,
-}
+enum AdNetwork { any, admob, appLovin, unity, facebook }
 
 extension AdNetworkExtension on AdNetwork {
   String get value => describeEnum(this);