+++ lib/src/utils/i_ad_id_manager.dart
@@ -1,38 +1,38 @@
 abstract class IAdIdManager {
   const IAdIdManager();
 
-  /// pass null if you wish to not to implementing unity ads
+  /// Pass null if you wish not to implement unity ads
   ///
   /// AppAdIds? get unityAdIds => null;
   AppAdIds? get unityAdIds;
 
-  /// pass null if you wish to not to implementing admob ads
+  /// Pass null if you wish not to implement admob ads
   ///
   /// AppAdIds? get admobAdIds => null;
   AppAdIds? get admobAdIds;
 
-  /// pass null if you wish to not to implementing facebook ads
-  ///
-  /// AppAdIds? get fbAdIds => null;
-  AppAdIds? get fbAdIds;
-
-  /// pass null if you wish to not to implementing appLovin ads
+  /// Pass null if you wish not to implement appLovin ads
   ///
   /// AppAdIds? get appLovinAdIds => null;
   AppAdIds? get appLovinAdIds;
+
+  /// Pass null if you wish not to implement facebook ads
+  ///
+  /// AppAdIds? get fbAdIds => null;
+  AppAdIds? get fbAdIds;
 }
 
 class AppAdIds {
   /// App Id should never be null, if there is no app id for a particular ad network, leave it empty
   final String appId;
 
-  /// if id is null then it will not be implemented
+  /// if id is null, it will not be implemented
   final String? interstitialId;
 
-  /// if id is null then it will not be implemented
+  /// if id is null, it will not be implemented
   final String? rewardedId;
 
-  /// if id is null then it will not be implemented
+  /// if id is null, it will not be implemented
   final String? bannerId;
 
   const AppAdIds({